CSP-673 validation and coverage added

+++ src/SFA.DAS.ApprenticeAccounts/Data/Models/MyApprenticeship.cs
@@ -9,7 +9,7 @@ public class MyApprenticeship : Entity
         // for Entity Framework
     }
 
-    public MyApprenticeship(Guid Id, Guid apprenticeId, int? uln, int? apprenticeshipId,string? employerName, DateTime? startDate, DateTime? endDate,int? trainingProviderId, string? trainingProviderName, string? trainingCode, string? standardUid )
+    public MyApprenticeship(Guid Id, Guid apprenticeId, long? uln, long? apprenticeshipId,string? employerName, DateTime? startDate, DateTime? endDate,long? trainingProviderId, string? trainingProviderName, string? trainingCode, string? standardUid )
     {
         this.Id = Id;
         ApprenticeId = apprenticeId;
@@ -27,12 +27,12 @@ public class MyApprenticeship : Entity
 
     public Guid Id { get; set; }
     public Guid ApprenticeId { get; set; }
-    public int? Uln { get; set; }
-    public int? ApprenticeshipId { get; set; }
+    public long? Uln { get; set; }
+    public long? ApprenticeshipId { get; set; }
     public string? EmployerName { get; set; }
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
-    public int? TrainingProviderId { get; set; }
+    public long? TrainingProviderId { get; set; }
     public string? TrainingProviderName { get; set; }
     public string? TrainingCode { get; set; }
